Added modules and function for creating base
