+++ homework4/yellow_pages/interface.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "phonebook.h" 
+
+int userInterface(int recordAmount, Record *currentBase);
