+++ homework4/yellow_pages/main.cpp
@@ -1,6 +1,10 @@
 #define _CRT_SECURE_NO_WARNINGS
 #include <stdio.h>
 #include "phonebook.h"
+#include "tests.h"
+#include "interface.h"
+
+const char baseFileName[] = "base.txt";
 
 int main()
 {
@@ -14,8 +18,9 @@ int main()
 		return 1;
 	}
 
-	struct Record base[baseSize]{};
+	auto base = createBase();
 	userInterface(readFile(base, baseFileName), base);
+	deleteBase(base);
 	return 0;
 }
