+++ homework4/yellow_pages/tests.cpp
@@ -2,14 +2,23 @@
 #include <string.h>
 
 const int realRecordAmount = 7;
+const char testBaseFileName[] = "test.txt";
+
+struct Record
+{
+	char name[nameLength];
+	char number[numberLength];
+};
 
 bool programTest()
 {
-	struct Record testBase[baseSize]{};
+	auto testBase = createBase();
+
 	int testRecordAmount = readFile(testBase, testBaseFileName);
 
 	if (testRecordAmount != realRecordAmount)
 	{
+		deleteBase(testBase);
 		return false;
 	}
 
@@ -26,6 +35,7 @@ bool programTest()
 
 	if (testRecordAmount != realRecordAmount + 1)
 	{
+		deleteBase(testBase);
 		return false;
 	}
 
@@ -33,16 +43,19 @@ bool programTest()
 	{
 		if (strcmp(testBase[searchName(testRecordAmount, testBase, testNumber1)].name, testName1) != 0)
 		{
+			deleteBase(testBase);
 			return false;
 		}
 
 		if (strcmp(testBase[searchName(testRecordAmount, testBase, testNumber2)].name, testName2) != 0)
 		{
+			deleteBase(testBase);
 			return false;
 		}
 	}
 	else
 	{
+		deleteBase(testBase);
 		return false;
 	}
 
@@ -50,28 +63,34 @@ bool programTest()
 	{
 		if (strcmp(testBase[searchNumber(testRecordAmount, testBase, testName1)].number, testNumber1) != 0)
 		{
+			deleteBase(testBase);
 			return false;
 		}
 
 		if (strcmp(testBase[searchNumber(testRecordAmount, testBase, testName2)].number, testNumber2) != 0)
 		{
+			deleteBase(testBase);
 			return false;
 		}
 	}
 	else
 	{
+		deleteBase(testBase);
 		return false;
 	}
 
 	if (searchName(testRecordAmount, testBase, testNumber3) != -1)
 	{
+		deleteBase(testBase);
 		return false;
 	}
 
 	if (searchNumber(testRecordAmount, testBase, testName3) != -1)
 	{
+		deleteBase(testBase);
 		return false;
 	}
 
+	deleteBase(testBase);
 	return true;
 }
