+++ homework4/yellow_pages/tests.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool programTest();
